refactor: folder structure
- move page components into pages folder
- move redux-connected components into containers folder
- components folder is for keeping unconnected components
