--- src/app/components/index.tsx
@@ -1,2 +1,0 @@
-// tslint:disable:export-name
-export {Header} from "./Header";
