--- src/app/containers/index.tsx
@@ -1,5 +1,0 @@
-/** Exporting Containers for Easier Imports */
-export {Html} from "./Html";
-export {App} from "./App";
-export {HomePage} from "./HomePage";
-export {AboutPage} from "./AboutPage";
