+++ src/app/pages/StarsPage.test.tsx
