+++ src/app/pages/StarsPage.tsx
